requirements.txt
fastapi==0.115.0
pydantic==2.9.2
pydantic_settings==2.5.2
uvicorn[standard]==0.30.6
strawberry-graphql[fastapi]==0.209.6
strawberry-graphql[cli]==0.209.6
gunicorn==23.0.0
aiologger==0.7.0
ujson==5.10.0
email-validator==2.0.0.post2
google-cloud-firestore==2.19.0
google-cloud-trace==1.13.5
google-cloud-logging==3.11.2
opentelemetry-api===1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-logging==0.48b0
